some changes and fixes
svn path=/trunk/kdemultimedia/k3b/; revision=136253

+++ src/data/k3bdatadoc.h
@@ -59,8 +59,8 @@ class K3bDataDoc : public K3bDoc
   void addView(K3bView* view);
 
   bool newDocument();	
-  long size() const;
-  int length() const;
+  unsigned long size() const;
+  unsigned long length() const;
 
   const QString& name() const { return m_name; }
 
@@ -198,7 +198,7 @@ class K3bDataDoc : public K3bDoc
   bool m_deleteImage;
   bool m_onlyCreateImage;
 
-  long m_size;
+  unsigned long m_size;
 		
   int m_whiteSpaceTreatment;
